Create String Replacement Example in C++.cpp
This C++ program demonstrates how to perform string replacement using the std::string class. The program replaces the substring "World" with "C++" in the string "Hello, World!".

The program uses the find method to locate the position of the substring to be replaced. If the substring is found, the replace method is used to replace it with the new substring. The result is then printed to the console.

This example provides a simple and clear way to understand how string manipulation can be done in C++, making it useful for beginners learning the language.